закончил рефакторинг
Разнес все по разным файлам почистил дублирование кода
