+++ src/index.js
@@ -4,20 +4,21 @@ import './sass/main.scss';
 import './js/refs/refs.js';
 
 //functions
-import './js/layout/fetch-by-Id';
-import './js/layout/render-by-params';
-import './js/layout/fetch-by-keyword';
-import './js/layout/week-movies';
 import './js/layout/local-storage';
 import './js/layout/render-by-template';
 import './js/layout/render-storage-array';
-import './js/layout/switchTheme.js';
-import './js/layout/render-genres-list';
+import './js/layout/transformation-genre';
 
 //header
 import './js/layout/header';
 import './js/layout/filter';
+import './js/layout/search-by-keyword';
+import './js/layout/switchTheme.js';
 //main
+import './js/layout/render-main-list';
+import './js/layout/fetch-by-Id';
+import './js/layout/week-movies';
+import './js/layout/render-genres-list';
 import './js/layout/pagination';
 
 //footer
@@ -36,7 +37,7 @@ import './js/components/no-scroll.js';
 import './js/api/api-service.js';
 
 //=============test==================
-import './js/test/test-api.js';
-import './js/test/test-video-player.js';
-import './js/test/test-modal.js';
+// import './js/test/test-api.js';
+// import './js/test/test-video-player.js';
+// import './js/test/test-modal.js';
 // import './js/test/test-notification.js';
