+++ src/js/api/api-service.js
@@ -9,7 +9,6 @@ export const getApiData = query => {
   return axios
     .get(`${query}&api_key=${API_KEY}&append_to_response=videos`)
     .then(response => {
-      console.log(response);
       hideSpiner();
       return response.data;
     })
