--- src/js/layout/fetch-by-keyword.js
@@ -1,3 +1,0 @@
-// import { refs } from '../refs/refs';
-
-// refs.searchFormEl.addEventListener('submit', onSearch);
